+++ screens/DashboardScreen.jsx
@@ -1,7 +1,7 @@
 import { ScrollView, View } from "react-native"
 import React from "react"
 import InboxWidget from "../components/InboxWidget"
-import ClientWidget from "../components/ClientWidget"
+import ClientWidget from "../components/ClientScreenComponents/ClientWidget"
 import ProtocolWidget from "../components/ProtocolScreenComponents/ProtocolWidget"
 
 const DashboardScreen = () => {
